Move default_pipelines nanobind module from frontend to catalyst Create DefaultPipelines folder both under mlib/include/Driver and mlir/lib/Driver. Move DefaultPipelines.h to mlir/include/Driver and DefaultPipelines.cpp to mlir/lib/Driver. Move default_pipelines nanobind module CMake code to mlir/include/Driver/CMakeLists.txt. Update target include directories of default_pipelines nanobind module to point to mlir/include/Driver. Remove linkage of default_pipelines nanobind module against CatalystCompilerDriver. Update setup.py to build extensions on different folders. Update setup.py to create catalyst.default_pipelines extension from mlir/lib/Driver/DefaultPipelines folder.
