+++ mlir/lib/Driver/DefaultPipelines/DefaultPipelines.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "Driver/DefaultPipelines.h"
+#include "DefaultPipelines/DefaultPipelines.h"
 
 #include <nanobind/nanobind.h>
 #include <nanobind/stl/string.h>
