+++ mlir/lib/Driver/Pipelines.cpp
@@ -16,9 +16,6 @@
 
 #include <memory>
 
-#include <fmt/core.h>
-#include <fmt/ranges.h>
-
 #include "mlir/Conversion/AffineToStandard/AffineToStandard.h"
 #include "mlir/Conversion/ArithToLLVM/ArithToLLVM.h"
 #include "mlir/Conversion/BufferizationToMemRef/BufferizationToMemRef.h"
@@ -44,7 +41,7 @@
 
 #include "Catalyst/IR/CatalystDialect.h"
 #include "Catalyst/Transforms/Passes.h"
-#include "Driver/DefaultPipelines.h"
+#include "Driver/DefaultPipelines/DefaultPipelines.h"
 #include "Driver/Pipelines.h"
 #include "Gradient/IR/GradientDialect.h"
 #include "Gradient/Transforms/Passes.h"
@@ -53,6 +50,9 @@
 #include "Quantum/IR/QuantumDialect.h"
 #include "Quantum/Transforms/Passes.h"
 #include "hlo-extensions/Transforms/Passes.h"
+
+#include <fmt/core.h>
+#include <fmt/ranges.h>
 
 using namespace mlir;
