+++ setup.py
@@ -288,7 +288,7 @@ class UnifiedBuildExt(build_ext):
         if "CMAKE_ARGS" in os.environ:
             configure_args += os.environ["CMAKE_ARGS"].split(" ")
 
-        build_temp = os.path.abspath(self.build_temp)
+        build_temp = os.path.join(os.path.abspath(self.build_temp), ext.name)
         os.makedirs(build_temp, exist_ok=True)
 
         build_args = ["--config", "Debug"] if debug else ["--config", "RelWithDebInfo"]
@@ -376,12 +376,12 @@ elif system_platform == "Darwin":
 
 project_root_dir = os.path.abspath(os.path.dirname(__file__))
 frontend_dir = os.path.join(project_root_dir, "frontend")
-driver_dir = os.path.join(project_root_dir, "mlir", "lib", "Driver")
+default_pipelines_dir = os.path.join(project_root_dir, "mlir", "lib", "Driver", "DefaultPipelines")
 
 ext_modules = [
     custom_calls_extension,
     CMakeExtension("catalyst.utils.wrapper", sourcedir=frontend_dir),
-    CMakeExtension("catalyst.default_pipelines", sourcedir=driver_dir),
+    CMakeExtension("catalyst.default_pipelines", sourcedir=default_pipelines_dir),
 ]
 
 options = {"bdist_wheel": {"py_limited_api": "cp312"}} if sys.hexversion >= 0x030C0000 else {}
